+++ weblate/trans/models/translation.py
@@ -21,6 +21,7 @@
 import codecs
 import os
 import tempfile
+from uuid import uuid4
 
 from django.core.cache import cache
 from django.core.exceptions import ValidationError
@@ -55,7 +56,7 @@ from weblate.trans.validators import validate_check_flags
 from weblate.utils.errors import report_error
 from weblate.utils.render import render_template
 from weblate.utils.site import get_site_url
-from weblate.utils.stats import TranslationStats
+from weblate.utils.stats import TranslationStats, ZeroStats
 
 
 class TranslationManager(models.Manager):
@@ -1075,3 +1076,19 @@ class Translation(models.Model, URLMixin, LoggerMixin):
             self.component.create_translations(request=request)
             self.__git_commit(request.user.get_author_name(), timezone.now())
             self.component.push_if_needed()
+
+
+class GhostTranslation:
+    """Ghost translation object used to show missing translations."""
+
+    is_ghost = True
+
+    def __init__(self, component, language):
+        self.component = component
+        self.language = language
+        self.stats = ZeroStats(component.source_translation.stats)
+        self.pk = uuid4().hex
+        self.is_source = False
+
+    def __str__(self):
+        return "{0} — {1}".format(self.component, self.language)